Floorplan: fix adjustment calculation
The base icon size is now stored in a global var. This value is used to
calculate the the adjustment.

Setting the percentage witdh of the icon and then using the value of the
width property in the same pass does not work, the returned value seems
to be something intermediate while the element is rendered or scaled or
something...