FP: add delayed call of fp_reposition_entities
I still have no clue why the initial load does not position the elements
at the correct positon. The values passed to the .offset method of the
items are the same on the first and subsequent runs..

If i add a breakpoint to fp_reposition_entities and then let it continue
the postioning gets much better and the items end up pretty near to
where they actually should be.

This change calls fp_reposition_entities after a timeout fo 1s after the
initial positioning. This works here for all my devices but may need
some tweaking..